porting 2022 oct 18 1250; in progress porting a_stb_truetype.rs

+++ src/stb_tt_buf.rs
@@ -1,11 +1,11 @@
 #![allow(non_camel_case_types)]
 
-use libc::{c_char, c_int};
+use libc::{c_char, c_int, size_t};
 
 // private structure
 #[derive(Default, Debug, Copy, Clone)]
 pub struct stbtt__buf {
     pub c_udata: *mut c_char,
-    pub cursor: c_int,
-    pub size: c_int,
+    pub cursor: size_t,
+    pub size: size_t,
 }
